task 1
Counts and prints to the console the number of categories in ul#categories, i.e. li.item elements.

For each li.item in the ul#categories list, find and print to the console the text of the element's title (the <h2> tag) and the number of items in the category (all its nested <li> s).